+++ it/example_test.go
@@ -979,3 +979,17 @@ func ExampleIPConstraint_DenyIP() {
 	// Output:
 	// violation: "This IP address is prohibited to use."
 }
+
+func ExampleHasNoSuspiciousCharacters_valid() {
+	err := validator.Validate(context.Background(), validation.String("alice", it.HasNoSuspiciousCharacters()))
+	fmt.Println(err)
+	// Output:
+	// <nil>
+}
+
+func ExampleHasNoSuspiciousCharacters_invalid() {
+	err := validator.Validate(context.Background(), validation.String("a\u200b", it.HasNoSuspiciousCharacters()))
+	fmt.Println(err)
+	// Output:
+	// violation: "Using invisible characters is not allowed."
+}